rename class to Gpt2
